+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 0.5.3
 **Bugfixes**
 - Fixed not being able to open a Grid that is 256 blocks away from the Controller
-- Change block hardness levels
+- Changed block hardness levels
 
 **Features**
 - Made the mod way less network intensive